+++ lib/src/extentions/list.dart
@@ -1,0 +1,4 @@
+extension WeeListExtension<T> on List<T> {
+  /// Ambil elemen terakhir sejumlah [count]
+  List<T> takeLast(int count) => skip(length - count).toList();
+}